backend/requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库和缓存
redis>=5.0.0
sqlite3  # Python内置，无需安装

# 环境变量管理
python-dotenv>=1.0.0

# AI和语言模型
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.20
openai>=1.0.0

# 向量数据库
chromadb>=0.4.0
langchain-community>=0.0.10

# 数据验证和序列化
pydantic>=2.0.0

# WebSocket支持
websockets>=12.0
